Revert "FAC-726: Move ports to the right package"
This reverts commit b13bf313df56e8bc87d8481dfe416ed1b11a8cc9.
